Break up ondemand_basictests into smaller executables
(Allows for faster compilation speeds)
